requirements.txt
Flask>=2.3.3
torch>=2.0.1
torchvision>=0.15.2
Pillow>=10.0.1
kornia>=0.6.13
matplotlib>=3.8.0
tqdm>=4.68.0